refactor: rounds to templates
